requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.36.23
botocore==1.36.23
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
dj-rest-auth==7.0.1
Django==4.2.19
django-allauth==65.4.1
django-crispy-forms==2.3
django-sslserver==0.22
django-storages==1.14.5
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
pillow==10.4.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==1.26.20
whitenoise==6.7.0
